refactor : feed page and PostList
