+++ src/components/posts/comments/CommentList.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from 'react'
-import '../../assets/styles/Comments/Comments.css'
+import '../../../assets/styles/Comments/Comments.css'
 import Comment from './Comment'
-import Textarea from '../textarea/Textarea'
+import Textarea from '../../textarea/Textarea'
+
 export interface IComment {
   id: number
   content: string
@@ -9,12 +10,16 @@ export interface IComment {
 interface ICommentList {
   setCommentNumber: React.Dispatch<React.SetStateAction<number>>
 }
+
 function CommentList(props: ICommentList) {
   const [comments, setComments] = useState<IComment[]>([])
   const [textComments, setTextComments] = useState<string>('')
   const { setCommentNumber } = props
 
-  useEffect(() => setCommentNumber(comments.length), [comments])
+  useEffect(
+    () => setCommentNumber(comments.length),
+    [comments, setCommentNumber]
+  )
   useEffect(() => {
     textComments &&
       setComments(previous => [
